Redirect from social to login activity, remove obsolete code.
